fix:正在使用的角色无法删除

+++ src/main/java/com/example/canary/sys/service/RoleServiceImpl.java
@@ -7,8 +7,10 @@ import com.example.canary.sys.entity.RolePO;
 import com.example.canary.sys.entity.RolePermissionPO;
 import com.example.canary.sys.entity.RoleQuery;
 import com.example.canary.sys.entity.RoleVO;
+import com.example.canary.sys.entity.UserRolePO;
 import com.example.canary.sys.repository.RolePermissionRepository;
 import com.example.canary.sys.repository.RoleRepository;
+import com.example.canary.sys.repository.UserRoleRepository;
 import com.example.canary.util.PageUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +19,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.CollectionUtils;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 /**
  * 角色
@@ -35,6 +36,9 @@ public class RoleServiceImpl implements RoleService {
     @Autowired
     private RolePermissionRepository rolePermissionRepository;
 
+    @Autowired
+    private UserRoleRepository userRoleRepository;
+
     /**
      * pages
      *
@@ -44,7 +48,7 @@ public class RoleServiceImpl implements RoleService {
     @Override
     public ResultEntity<IPage<RoleVO>> pagesRole(RoleQuery query) {
         IPage<RolePO> pagePo = roleRepository.selectPagePo(query);
-        List<RoleVO> records = pagePo.getRecords().stream().map(RoleVO::new).collect(Collectors.toList());
+        List<RoleVO> records = pagePo.getRecords().stream().map(RoleVO::new).toList();
         IPage<RoleVO> pageVo = PageUtils.convertToVo(pagePo, records);
         return ResultEntity.success(pageVo);
     }
@@ -103,6 +107,11 @@ public class RoleServiceImpl implements RoleService {
     @SuppressWarnings("rawtypes")
     @Transactional(rollbackFor = Exception.class)
     public ResultEntity deleteRole(String roleId) {
+        // 查询角色使用情况
+        List<UserRolePO> userRoles = userRoleRepository.selectByRoleId(roleId);
+        if (!CollectionUtils.isEmpty(userRoles)) {
+            return ResultEntity.fail("正在使用的角色无法删除");
+        }
         // delete role
         roleRepository.deleteById(roleId);
         // delete relation
